backend/requirements.txt
fastapi==0.115.7
uvicorn==0.34.0
python-dotenv==1.0.1
requests==2.32.3
yfinance==0.2.52
openai==1.60.1
azure-cognitiveservices-language-textanalytics==0.2.2
pandas==2.2.0
numpy==1.26.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
sqlalchemy==2.0.27
pydantic==2.6.1
pydantic-settings==2.1.0 